--- android/app/src/main/java/com/zzang/chongdae/data/remote/source/CommentDetailDataSource.kt
@@ -1,16 +1,0 @@
-package com.zzang.chongdae.data.remote.source
-
-import com.zzang.chongdae.data.remote.dto.request.CommentRequest
-import com.zzang.chongdae.data.remote.dto.response.CommentsResponse
-import com.zzang.chongdae.data.remote.dto.response.MeetingsResponse
-
-interface CommentDetailDataSource {
-    suspend fun getMeetings(offeringId: Long): Result<MeetingsResponse>
-
-    suspend fun saveComment(commentRequest: CommentRequest): Result<Unit>
-
-    suspend fun fetchComments(
-        offeringId: Long,
-        memberId: Long,
-    ): Result<CommentsResponse>
-}
