--- android/app/src/main/java/com/zzang/chongdae/data/remote/source/CommentRoomsDataSource.kt
@@ -1,7 +1,0 @@
-package com.zzang.chongdae.data.remote.source
-
-import com.zzang.chongdae.data.remote.dto.response.CommentRoomsResponse
-
-interface CommentRoomsDataSource {
-    suspend fun fetchCommentRooms(memberId: Long): Result<CommentRoomsResponse>
-}
