+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/CommentRoomsDataSourceImpl.kt
@@ -1,8 +1,8 @@
-package com.zzang.chongdae.data.remote.source.impl
+package com.zzang.chongdae.data.remote.source
 
 import com.zzang.chongdae.data.remote.api.CommentsApiService
 import com.zzang.chongdae.data.remote.dto.response.CommentRoomsResponse
-import com.zzang.chongdae.data.remote.source.CommentRoomsDataSource
+import com.zzang.chongdae.data.source.CommentRoomsDataSource
 
 class CommentRoomsDataSourceImpl(
     private val commentsApiService: CommentsApiService,
